Initial commit of a Natron-specific OpenColorIO congifuration.
This commit attempts to build a clean configuration for imagers
using the Natron compositor. It includes a number of transformations
relevant to those working with motion picture imaging formats, as
well as using generic families to organize the transforms. These
families may be leveraged in Natron's UI in the future to streamline
colourspace related interface elements.

Incorporates feedback from @devernay.